02_Intermediate_projects/018_zenflow/pyproject.toml
[project]
name = "zenflow"
version = "0.1.0"
dependencies = [
    "click",
    "pyyaml",
]

[project.scripts]
flow = "flow.cli:entry_point"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"